backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
httpx
apscheduler
unidecode
alembic
gunicorn
pytest